requirements.txt
sds011==0.2.0
ipython
flask
#redis
APScheduler==3.7.0
flask-cors
python-aqi==0.6.1

